+++ CompanionAppService/CompanionAppService.cs
@@ -225,7 +225,7 @@ namespace EddiCompanionAppService
             CurrentState = State.NEEDS_LOGIN;
         }
 
-        public Profile Profile()
+        public Profile Profile(bool forceRefresh = false)
         {
             Logging.Debug("Entered");
             if (CurrentState != State.READY)
@@ -235,7 +235,7 @@ namespace EddiCompanionAppService
                 Logging.Debug("Leaving");
                 throw new EliteDangerousCompanionAppIllegalStateException("Service in incorrect state to provide profile (" + CurrentState + ")");
             }
-            if (cachedProfileExpires > DateTime.Now)
+            if ((!forceRefresh) && cachedProfileExpires > DateTime.Now)
             {
                 // return the cached version
                 Logging.Debug("Returning cached profile");
@@ -567,18 +567,18 @@ namespace EddiCompanionAppService
         private static void AugmentCmdrInfo(Commander cmdr)
         {
             Logging.Debug("Entered");
-            if (cmdr != null)
-            {
-                CommanderConfiguration cmdrConfiguration = CommanderConfiguration.FromFile();
-                if (cmdrConfiguration.PhoneticName == null || cmdrConfiguration.PhoneticName.Trim().Length == 0)
-                {
-                    cmdr.phoneticname = null;
-                }
-                else
-                {
-                    cmdr.phoneticname = cmdrConfiguration.PhoneticName;
-                }
-            }
+            //if (cmdr != null)
+            //{
+            //    CommanderConfiguration cmdrConfiguration = CommanderConfiguration.FromFile();
+            //    if (cmdrConfiguration.PhoneticName == null || cmdrConfiguration.PhoneticName.Trim().Length == 0)
+            //    {
+            //        cmdr.phoneticname = null;
+            //    }
+            //    else
+            //    {
+            //        cmdr.phoneticname = cmdrConfiguration.PhoneticName;
+            //    }
+            //}
             Logging.Debug("Leaving");
         }