+++ ChangeLog.md
@@ -1,3 +1,14 @@
+#2.1.0-b2
+  * Add 'Engineer' government type for engineer bases
+  * Enable update and outdate messages
+  * Ensure that VoiceAttack decimal values are not written as integers
+  * Add 'SetState' Cottle function.  Details on how to use this are in the SpeechResponder documentation
+  * Allow speech responder to not speak when subtitles are written
+  * Retry companion API profile call if returned information is outdated.  This should help to avoid situations where outfitting and market data is out-of-date
+  * Add event when being scanned for cargo by an NPC
+  * Add event when being attacked by an NPC
+  * Add event when being interdicted by an NPC
+
 #2.1.0-b1
   * Fix material name for Cadmium
   * Add ability to set state variables from VoiceAttack.  Details on how to use this are in the VoiceAttack documentation